+++ NpgsqlRest/DefaultNameConverter.cs
@@ -1,10 +1,10 @@
 ﻿namespace NpgsqlRest;
 
-internal static class DefaultNameConverter
+public static class DefaultNameConverter
 {
     private static readonly string[] separator = ["_"];
 
-    internal static string? ConvertToCamelCase(string? value)
+    public static string? ConvertToCamelCase(string? value)
     {
         if (value is null)
         {
@@ -18,6 +18,7 @@ internal static class DefaultNameConverter
             .Split(separator, StringSplitOptions.RemoveEmptyEntries)
             .Select((s, i) => 
                 string.Concat(i == 0 ? char.ToLowerInvariant(s[0]) : char.ToUpperInvariant(s[0]), s[1..]))
-            .Aggregate(string.Empty, string.Concat);
+            .Aggregate(string.Empty, string.Concat)
+            .Trim('"');
     }
 }